+++ jmx/jmx-api/src/main/java/uk/gov/justice/services/jmx/api/mbean/InvalidHandlerMethodException.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.services.jmx.command;
+package uk.gov.justice.services.jmx.api.mbean;
 
 public class InvalidHandlerMethodException extends RuntimeException  {