+++ jmx/jmx-command-handling/src/main/java/uk/gov/justice/services/jmx/command/HandlerMethodValidator.java
@@ -3,6 +3,7 @@ package uk.gov.justice.services.jmx.command;
 import static java.lang.String.format;
 import static java.lang.reflect.Modifier.isPublic;
 
+import uk.gov.justice.services.jmx.api.mbean.InvalidHandlerMethodException;
 import uk.gov.justice.services.jmx.api.command.SystemCommand;
 
 import java.lang.reflect.Method;